Updated CMakeLists.txt for linking filesystem library conditionally
CMakeLists.txt now checks for the compiler ID. If one is not using GNU compiler, it gives a fatal error and exits. I suppose the support can be extended to LLVM Clang compilers, but I am not sure how to do that. CMake now links stdc++fs if the GNU C++ compiler has version less than 7.0 (in which case, <experimental/filesystem> is being used), otherwise it doesn't link stdc++fs.