Removed inline from the parse_id function in helpers.h
helpers.h contained the parse_id function's definition, rather than the declaration, and hence, when it was being included at multiple places, it led to failure of building the program. Now the definition (placed in a new helpers.cpp file) has been separated from the declaration, so it builds successfully without the inline qualifier.