Removed a few boost dependencies (#10)
Replace boost::array, boost::foo_pointer, boost::filesystem, boost::lexical_cast with the appropriate std:: replacements.

* Removed dependency on boost::array

Simply replaced the usage of boost::array by std::array

* Removed dependency on boost::array and boost::shared_ptr

Replaced the usage of boost::array by std::array and of boost::shared_ptr with std::shared_ptr (which can be found in the stanrd C++ <memory> header)

* Removed the dependency on boost::filesystem

Simply replaced the usage of boost::filesystem by std::filesystem. Tested it on my machine, works as expected.

* Removed dependency on boost::array and boost::shared_ptr

* Removed the dependency on boost::lexical_cast

C++11 and above provides std::to_string with appropriate overloads to convert arithmetic types to std::string. However, it doesn't preserve accuracy when used to convert floating-point variables. But in the project, all the cases are regarding integer types, so it should work perfectly fine.

* Removed outdated TODO

* Added check for filesystem header

Checks whether <filesystem> header exists, if not then checks for <experimental/filesystem>. If none of these are present, it produces an error. [Reference](https://stackoverflow.com/questions/53365538/how-to-determine-whether-to-use-filesystem-or-experimental-filesystem).

* Added linker flag for std::experimental::filesystem

* Updated CMakeLists.txt for linking filesystem library conditionally

CMakeLists.txt now checks for the compiler ID. If one is not using GNU compiler, it gives a fatal error and exits. I suppose the support can be extended to LLVM Clang compilers, but I am not sure how to do that. CMake now links stdc++fs if the GNU C++ compiler has version less than 7.0 (in which case, <experimental/filesystem> is being used), otherwise it doesn't link stdc++fs.

* Added TODO

* Revert "Added TODO"

This reverts commit ffe997c02ba166615171254263314b95df4c3f71.

TODO not relevant anymore

* Revert "Updated CMakeLists.txt for linking filesystem library conditionally"

This reverts commit 8d4072f9920b7ef6278ac2c50fab38cb4d8c6e1b.

Breaks the build process

We need to find a better way to check when to link stdc++fs and when not to. The current way breaks the build process.

* Added relevant TODO

* Removed unnecessary TODO